QOL AND PERFORMANCE IMPROVEMENTS
Script now creates new blank files for templates, blacklists, and prompt logs if they are moved or deleted. This allows you to delete templates and blacklists, and create new ones from a blank slate without editing the .json if you don't like the default templates and blacklists.

Logs are now saved and shown in reverse order, so the newest logged prompt is always at the top.